[JCN-349-create-lambda-functions-with-janis-cli] - unify schema templates and functions
