+++ lib/commands/create/function.js
@@ -1,7 +1,6 @@
 'use strict';
 
-const functionWithPayloadSchemaTemplate = require('../../templates/function-with-payload-schema');
-const functionWithoutPayloadSchemaTemplate = require('../../templates/function-without-payload-schema');
+const functionSchemaTemplate = require('../../templates/function-schema');
 const functionSchemaFs = require('../../fs/function-schema');
 
 const functionSourceTemplate = require('../../templates/function-source');
@@ -86,7 +85,7 @@ module.exports.handler = async argv => {
 	const promises = [];
 
 	// Function Schema
-	const schemaContent = hasPayload ? functionWithPayloadSchemaTemplate({ functionName }) : functionWithoutPayloadSchemaTemplate({ functionName });
+	const schemaContent = functionSchemaTemplate({ functionName, hasPayload });
 	promises.push(functionSchemaFs.writeSchema(functionPath, functionName, schemaContent));
 
 	// Function Source
