+++ lib/templates/function-schema.js
@@ -3,18 +3,21 @@
 const apiSchemaTemplate = require('./api-schema');
 
 module.exports = ({
-	functionName
+	functionName,
+	hasPayload
 }) => apiSchemaTemplate({
 	path: `/${functionName}`,
 	method: 'post',
 	operationId: `${functionName}`,
 	tags: ['Functions'],
 	summary: `${functionName}`,
-	requestBody: {
-		description: `${functionName}`,
-		required: true,
-		content: {
-			'application/json': {
+	...hasPayload && {
+		requestBody: {
+			description: `${functionName}`,
+			required: true,
+			content: {
+				'application/json': {
+				}
 			}
 		}
 	},
