--- lib/templates/function-without-payload-schema.js
@@ -1,18 +1,0 @@
-'use strict';
-
-const apiSchemaTemplate = require('./api-schema');
-
-module.exports = ({
-	functionName
-}) => apiSchemaTemplate({
-	path: `/${functionName}`,
-	method: 'post',
-	operationId: `${functionName}`,
-	tags: ['Functions'],
-	summary: `${functionName}`,
-	responses: {
-		200: {
-			$ref: '#/components/responses/GenericSuccess'
-		}
-	}
-});
